requirements.txt
validators
customtkinter
Pillow
tkcalendar
ttkwidgets


